+++ setup.py
@@ -1,7 +1,21 @@
 from setuptools import setup, find_packages
 from os.path import abspath, dirname, join as ospjoin
+import re
 
 here = abspath(dirname(__file__))
+
+
+def find_version(filename):
+    with open(filename, 'r') as f:
+        version_file = f.read()
+    version_match = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]",
+                              version_file, re.M)
+    if version_match:
+        return version_match.group(1)
+    raise RuntimeError("Unable to find version string.")
+
+
+project_version = find_version(ospjoin(here, 'fastnml', '__init__.py'))
 
 # Get the long description from the README file
 with open(ospjoin(here, 'README.md'), encoding='utf-8') as f:
@@ -9,7 +23,7 @@ with open(ospjoin(here, 'README.md'), encoding='utf-8') as f:
 
 setup(
     name='fastnml',
-    version='0.1.0',
+    version=project_version,
     description='A simple fast namelist parser',
     long_description=long_description,
     long_description_content_type='text/markdown',
